Initial release of No-rebase.github.io page
Feel free to fork and modify as you please.

Bugs reports and PRs are appreciated, at

https://github.com/no-rebase/No-rebase.github.io/issues

Here are some examples of trailers declining consent:
(adding the name is optional; it enables a more personalized message,
ie, adding ?drok to the URL results in "drok does not consent to
Rebasing their PR" with a hovercard)

See the consent page for instructions on how to automate adding the
trailer to commits.

Author-Rebase-Consent: https://No-rebase.github.io?drok